chore: Configure native handlers. + Fix serialization of AuthenticationToken.
